toolchain: refactor
In this commit, the prebuilt gcc toolchain's checksums are stored
in a json file, rather than hard-coded in the script. Furthermore,
some functions are defined to make it possible for GCC toolchain
to uses the NDK sysroot of CURRENT TOOLCHAIN rather the one of r17c.
Hopefully this will be helpful to enable the C11 support for the
custom GCC toolchain, especially the `<threads.h>` header, which
is needed by `scipy`.