addpkg(main/yc): Yao scripting language and Rig build system
More information about Yao and Rig is at https://rigbuild.dev/ .

Yao and Rig are part of a monorepo, so this package includes both.

Since Rig is a build system, it is its own build system, so this package
requires a host build for bootstrap.

Both are currently in alpha, but a Termux user requested that I make this
package. That said, I understand if this package is not wanted,
especially since this package is under the SSPL and another non-Open
Source license, which may cause problems with parsing the license
identifier.

If accepted, I intend to maintain this package.

Signed-off-by: Gavin D. Howard <gavin@yzena.com>