dosemu2: add gfx support
It turned out dosemu2 was finding libx11 during build and was
enabling the basic GUI support, even if I didn't ask it for that.

This patch enables the official GFX support via libsdl2.
It disables outdated X support which currently needs libandroid-shmem
and its not being linked in.
`sdl2 | sdl2-compat` is used for appropriate libsdl2 provider.

Related: #1833